add ExecutionPlanHelper and fix naming
ONE-DCO-1.0-Signed-off-by: Artem Balyshev a.balyshev@partner.samsung.com
