+++ compiler/circle-execution-plan/src/ExecutionPlanner.h
@@ -17,8 +17,6 @@
 #ifndef CIRCLE_EXECUTION_PLANNER_H
 #define CIRCLE_EXECUTION_PLANNER_H
 
-#include <luci/IR/Module.h>
-#include "NodeOrderInform.h"
 #include "ExecutionOrderEstimator.h"
 #include <luci/Plan/CircleNodeExecutionPlan.h>
 
@@ -68,9 +66,9 @@ public:
   // memory offsets for all nodes in _graph.
   // This plan writes in nodes annotation information with help of CircleNodeExecutionPlan class.
 
-  void make_default_execution_plan();
+  void make_rpo_execution_plan();
 
-  void make_estimated_execution_plan();
+  void make_exhaustive_search_execution_plan();
 
   // Method change planning mode:
   // is_null_consts = true - constants are no longer taken into account when planning
@@ -85,10 +83,6 @@ public:
 
 private:
   void make_execution_plan();
-
-  // Method gets default execution order plan and saves it in _ordered_nodes vector.
-  // There can be different variants of execution order and this method provides main one.
-  void get_default_execution_order_plan();
 
   // Method provides nodes with usage interval information.
   void get_usage_interval();
